Move `SessionRequest` into the `data` package
